feat(BlueTooth): 完成蓝牙DEMO，补充蓝牙文档
