+++ src/BluetoothDemo/SearchPage.less
@@ -1,0 +1,130 @@
+.bluetooth-search-container .btn.btn-primary {
+	border: 1px solid #0052d9;
+	background: #0052D9;
+	color: #fff;
+}
+
+.bluetooth-search-container .btn {
+	position: relative;
+	font-size: 16px;
+	width: 100%;
+	height: 45px;
+	line-height: 45px;
+	text-align: center;
+	background-color: #fff;
+	border-radius: 0;
+	border: 1px solid #DDDDDD;
+	color: #0052D9;
+}
+
+.bluetooth-search-container button {
+	position: relative;
+	display: block;
+	-webkit-box-sizing: border-box;
+	box-sizing: border-box;
+	margin: 0;
+	padding: 0;
+	-webkit-tap-highlight-color: transparent;
+	color: inherit;
+	background-color: inherit;
+}
+
+.bluetooth-search-container button::after {
+	width: 200%;
+	height: 200%;
+	position: absolute;
+	top: 0;
+	left: 0;
+	border: none;
+	-webkit-transform: scale(.5);
+	transform: scale(.5);
+	-webkit-transform-origin: 0 0;
+	transform-origin: 0 0;
+	box-sizing: border-box;
+	border-radius: 10px;
+
+}
+
+.bluetooth-search-container {
+	background: rgb(99, 113, 132);
+	padding: 25px
+}
+
+.bluetooth-search-item-list {
+	margin-top: 30px;
+	width: 100%;
+
+	.error {
+		color: #ff584c;
+	}
+
+	.searching-txt {
+		text-align: center;
+		line-height: 46 rpx;
+	}
+
+	.search-result {
+		width: 100%;
+		height: 100%;
+		padding: 24px 30px;
+		overflow: hidden;
+		box-sizing: border-box;
+
+		.search-title {
+			font-weight: bold;
+			font-size: 16px;
+			color: #000000;
+		}
+
+		.search-body {
+			height: 100%;
+			margin-top: 12px;
+			overflow-x: hidden;
+			overflow-y: scroll;
+		}
+	}
+
+	.search-row {
+		display: flex;
+		flex-wrap: wrap;
+
+		.item-name {
+			display: inline-block;
+			font-size: 26px;
+			color: #000000;
+			max-width: 80%;
+			margin-left: 4px;
+			line-height: 48px;
+			min-height: 48px;
+			vertical-align: top;
+		}
+
+		.search-item {
+			position: relative;
+			height: 90px;
+			width: 100%;
+			margin-bottom: 20px;
+			box-sizing: border-box;
+			background-color: rgba(0, 0, 0, 0.10);
+			border-radius: 12px;
+			padding: 10px;
+
+
+			.item-name {
+				display: inline-block;
+				font-size: 14px;
+				color: #ffffff;
+				line-height: 70px;
+			}
+
+			.link-btn {
+				position: absolute;
+				right: 30px;
+				color: #7cb342;
+				top: 40%;
+				font-size: 18px;
+				font-weight: bold;
+			}
+		}
+	}
+}
