+++ src/BluetoothDemo/index.js
@@ -1,0 +1,3 @@
+export * from './SearchPage';
+export * from './PanelPage';
+export * from './DemoBluetoothDeviceAdapter';
