+++ src/DevicePanel/DevicePanel.jsx
@@ -1,11 +1,8 @@
 import React, { useEffect, useState, useReducer } from 'react';
-// import sdk from 'qcloud-iotexplorer-h5-panel-sdk';
-import sdk from '../../../iot-explorer-h5-panel/sdk/src';
+import sdk from 'qcloud-iotexplorer-h5-panel-sdk';
 import classNames from 'classnames';
 import { NumberPanelControl } from '../components/NumberPanelControl';
 import { EnumPanelControl } from '../components/EnumPanelControl';
-
-console.log('sdk', sdk);
 
 function reducer(state, action) {
   const { type, payload } = action;
